Move test from viewbox-test
