+++ test/test_viewbox.py
@@ -263,22 +263,3 @@ class TestElementViewbox(unittest.TestCase):
         q = list(m.elements())
         self.assertEqual(len(q), 1)
         self.assertEqual(Matrix(m.viewbox_transform), Matrix.translate(1, 1))
-
-    def test_issue_157(self):
-        q = io.StringIO(u'''<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<svg>
-  <g id="layer1">
-    <text
-       style="font-size:18px;line-height:1.25;font-family:sans-serif;stroke-width:0.264583"
-       id="textobject"><tspan
-         id="tspanobject"
-         x="0"
-         y="0">Test</tspan></text>
-  </g>
-</svg>
-        ''')
-        m = SVG.parse(q)
-        q = list(m.elements())
-        self.assertIsNotNone(q[1].id) # Group
-        self.assertIsNotNone(q[2].id) # Text
-        self.assertIsNotNone(q[3].id) # TSpan
